+++ src/index.ts
@@ -1,3 +1,3 @@
-import Notification from './Notification';
+import useNotification from './useNotification';
 
-export default Notification;
+export { useNotification };
