+++ src_bac/index.ts
@@ -1,0 +1,3 @@
+import Notification from './Notification';
+
+export default Notification;
